Created Cv component
I am trying to figure out how to work with pdf on a webpage. the easiest so far is the iframe, but i dont like how it appears on the screen and how little manipulation i get to change the appearance of the pdf and the size of it. on going